deb: Increase darkest possible auto-brightness adjustment
A value of 1 is too low and results in backlight turning off when
adjusted even lower with CABC on. Increase it to 2.

Ideally we do not want this overlay change since it increases
the minimum brightness quite noticeably in dark ambient conditions.

However, we'll just merge this for now until a better fix is found.

Change-Id: I8eee6be1cffb7efe5c807dd235bc7d3b4624887d